// src/store/user/type.ts
export interface UserState {
  uid?: number;
  username?: string;
  email?: string;
  photo?: string;
  gender?: string;
  coin?: number;
  album_favorites?: number;
  level_name?: string;
  level?: number;
  nextLevelExp?: number;
  exp?: string;
  expPercent?: number;
  album_favorites_max?: number;
}